노멀투표/상세 분리 + 투표 백엔드 json데이터 끝

+++ backend/src/main/java/org/usyj/makgora/profile/service/ProfileActivityService.java
@@ -171,7 +171,7 @@ public List<RecentCommunityActivityResponse> getRecentCommunityActivities(Intege
 
             result.add(
                     RecentVoteActivityResponse.builder()
-                            .voteUserId(vu.getId())
+                            .voteUserId(vu.getId().intValue())
                             .voteId(vote.getId())
                             .voteTitle(vote.getTitle())
                             .issueTitle(issueTitle)
